backend/requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14
django-cors-headers>=4.3
channels>=4.0
channels-redis>=4.2
daphne>=4.0
requests>=2.31
httpx>=0.27
psycopg2-binary>=2.9
python-dotenv>=1.0
